chore: feature support section, tweak test config
