+++ web/pages/index.tsx
@@ -1,24 +1,22 @@
 import React, { useRef } from 'react';
-import { HeroBanner } from 'molecules/HeroBanner';
 import {
-  Title,
-  Text,
-  TextInput,
+  Button,
   Card,
   Center,
-  useMantineTheme,
-  Button,
   Group,
+  Text,
+  TextInput,
+  Title,
   createStyles,
-  Container,
-  Table,
+  useMantineTheme,
 } from '@mantine/core';
-import { Features, Faq } from 'molecules/sections';
-import { useRouter } from 'next/router';
-import { GRAB_PATH, OVERVIEW_DOCS_PATH } from '@shared/config/paths';
 import { useMediaQuery } from '@mantine/hooks';
+import { GRAB_PATH, OVERVIEW_DOCS_PATH } from '@shared/config/paths';
+import { HeroBanner } from 'molecules/HeroBanner';
+import { Faq, Features } from 'molecules/sections';
+import { FeaturesSupport } from 'molecules/sections/FeaturesSupport';
 import { Tools } from 'molecules/sections/Tools';
-import { SectionTitle } from 'molecules/sections/SectionTitle';
+import { useRouter } from 'next/router';
 
 const useStyles = createStyles((theme) => ({
   control: {
@@ -37,64 +35,6 @@ const useStyles = createStyles((theme) => ({
     },
   },
 }));
-
-type SupportStatus = '✅ stable' | '📝 planned' | '🧪 experimental';
-type FeatureEntry = {
-  name: string;
-  web: SupportStatus;
-  cli: SupportStatus;
-  vscode: SupportStatus;
-};
-
-const items: Array<FeatureEntry> = [
-  {
-    name: 'drop string',
-    web: '✅ stable',
-    cli: '✅ stable',
-    vscode: '📝 planned',
-  },
-  {
-    name: 'drop file',
-    web: '✅ stable',
-    cli: '✅ stable',
-    vscode: '📝 planned',
-  },
-  {
-    name: 'grab string',
-    web: '✅ stable',
-    cli: '✅ stable',
-    vscode: '📝 planned',
-  },
-  {
-    name: 'grab file',
-    web: '✅ stable',
-    cli: '✅ stable',
-    vscode: '📝 planned',
-  },
-  {
-    name: 'local vaults',
-    web: '📝 planned',
-    cli: '🧪 experimental',
-    vscode: '📝 planned',
-  },
-  {
-    name: 'cloud vaults',
-    web: '📝 planned',
-    cli: '📝 planned',
-    vscode: '📝 planned',
-  },
-];
-
-const FeatureRow = ({ item }: { item: FeatureEntry }) => {
-  return (
-    <tr key={item.name}>
-      <td>{item.name}</td>
-      <td>{item.web}</td>
-      <td>{item.cli}</td>
-      {/* <td>{item.vscode}</td> */}
-    </tr>
-  );
-};
 
 const Home = () => {
   const router = useRouter();
@@ -172,34 +112,7 @@ const Home = () => {
       </Center>
       <Faq />
       <Tools />
-      <Container
-        size={'lg'}
-        py={'md'}
-        px={0}
-        mb={theme.spacing.xl * 2}
-      >
-        <SectionTitle
-          label={'Features Support'}
-          id={'features-section'}
-        />
-        <Card maw={'650px'}>
-          <Table fontSize={'lg'}>
-            <thead>
-              <tr>
-                <th>Feature</th>
-                <th>Web Application</th>
-                <th>CLI</th>
-                {/* <th>VSCode Extension</th> */}
-              </tr>
-            </thead>
-            <tbody>
-              {items.map((item, index) => (
-                <FeatureRow key={index} item={item} />
-              ))}
-            </tbody>
-          </Table>
-        </Card>
-      </Container>
+      <FeaturesSupport />
       <Center>
         <Button
           className={classes.control}
