+++ web/tests/e2e/util.ts
@@ -81,6 +81,7 @@ export const createContextForBrowser = async (
       sameSite: 'Strict',
       url: baseURL,
       httpOnly: false,
+      secure: !isLocal,
     },
   ]);
